Refactored the Attack a bit
The validators have been moved to the parent class Order (even when they
are specific to a single type of Order, this seemed cleaner), as have
the common validation.

Note, however, that Order being an kind-of abstract class (it's not one
in the Rails STI [`abstract_class`](http://api.rubyonrails.org/classes/ActiveRecord/Inheritance/ClassMethods.html#abstract_class)
meaning, but that's another story), I've kept the tests duplicated in
the `MoveTest` and `AttackTest` suites.