First step toward a turn resolution system
We need a way to _resolve_ a turn – that is, apply all the orders
(possibly cancelling some that would be invalid or nulled) and
generate a new Board. I feel like this responsibility fits the Turn model
nicely – we _could_ give it to an even more specialized object such as
a TurnResolver or even a TurnResolutionService (please, no!), but this
seems overkill, at least for now.