Move resolution, second approach
I feel like the logic for a move resolution (i.e. updating the `Board`
object by replacing the relevant `Zone` data objects) should be in its
own method, decoupled from a given `Turn` instance. Let's try to move it
to the `Turn` class. So now the `Turn` object (the _class_ itself) is
responsible for updating a turn's board according to a move.