`Board#move` replaces `Board#apply_move`
It's not just a change of name: instead of a `Move` object, the method
expects only the relevant details of the move: how many units, from
where, to where.

By having the method expect simple data instead of a complete object,
we reduce the coupling – now the `Board` object is oblivious to what
makes a valid "move" – it only cares about shuffling units around.

This means that it will **not** be up to the board and its `#move` method
to ensure that the move is valid, allowed, etc. – this kind of mechanic
will move back to the `Turn` object, which is the actual "move manager".

`Board` is back to being mostly a data container, with a somewhat complex
API to manip said data; the rules logic are handled by other objects, such
as `Order`, `Move`, and especially `Turn`.