`Board#move` knows when to add or replace units
If when units are moved from a zone to another, they should replace
the units on the target zone, unless these units belong to the same player.

This is a bit of logic, but I don't consider it part of the _rules_ logic:
it's more the act of putting pieces on the board or removing them.