Pushing the "colliding orders detection" logic down to the DB
I like this better. Not for the potential efficiency gain of replacing
sorting work on the application layer by querying work on the DB layer,
but because it simplifies the resolution code – now the resolution
has little setup, it's mostly a readable procedure: cancel the colliding
moves, carry out the moves to be carried out, save the new turn, etc.

The new tests are a PITA, though, because they'd need a lot of setup –
so either tedious boilerplate or new fixtures.