Added tests for `Turn.orders` extensions
I started with 2 distinct tests, one for `#colliding` and one for
`#to_carry_out`, but eventually decided to merge the two. It is a bit
unusual, and could bite me later if the definition of these scopes
diverge, but at the moment one is really the negation of the other,
so paying only once the setup price for the test seems sound.