Fixed the `Order.colliding` bug
The previous implementation was too simple, and buggy: an order could
be counted as colliding if it has the same target as with _any_
colliding order, even in a different turn.

(That will teach me not to test scopes…)

Not that I'm a bit on the fence on this solution. Not so much because
of the possible performance issue (as said in the code comments: we're
not there yet), but because it required to write raw SQL and move away
from the nice abstractions of ActiveRecord. I don't know if it's a legit
"power user" case, or just that I don't know AR well enough to enjoy its
"conceptual compression". I basically wrote the SQL that I wanted, and
then backtracked to have as many AR stuff as possible.

Also note that I tried to use NOT EXISTS instead of a the "LEFT JOIN +
IS NULL" combo, but couldn't make it work. (Obviously, I'm in the
demographics for conceptual compression.)

As a bonus, I cleaned up the code a bit and introduced my frivolous
darling, `Module.concerning`.