Resolution of successful attacks
Note that I've updated the rules a bit: leaving a unit behind is
still mandatory, but a player can decide to go "all-in" and attack
_without_ keeping at least one unit on its origin no matter how many
units they have. (Before, it was only possible when attacking with
a single unit.)

In any case, a successful "all-in" attack still results in a empty target
which _cannot_ be occupied. This makes movement computation easier, and
could provide an interesting strategy element.