Move attack resolution to Attack + no board change
This is a big change to the app's architecture, in two parts.

First, the responsibility of _resolving an attack_ is moved to the
`Attack` object itself. This makes a lot of sense, and was hinted at
by the amount of _feature envy_ for attack that `Turn#resolve!` was
displaying. (Actually, it still is displaying it, since the code
hasn't been removed yet.) a `Turn` object should have _some_ responsibility
in the resolving of _itself_ – by cancelling conflicting orders, for
example – but when it comes to the resolution of the individual orders,
the latter should be in charge.

However, there was a reason for keeping everything under the umbrella
of the `Turn` objects, and that was the need to _update the board
for the turn that follows the resolution_.

Resolving a turn means updating the board; and since each turn has
a brand new `Board` object, updated from the resolved orders of the
previous turn, it also meant _creating a new board, and applying
the results of the resolved actions to it_. Upon resolution, a turn is
in charge of preparing the board for its successor; therefore, handling
the resolutions of the orders was easier – otherwise, the new board
would have had to be passed to the action and used during the resolution…

Now, the second change introduced by this commit is that _order resolution
don't update any board_. Instead, resolving an order returns a
_board update_ – a series of changes that _should_ eventually by applied
to the board.

Having a list of updates to be applied, instead of applying changes
directly, will allow us to resolve potential conflicts. For example,
how should we handle a player attacking a zone that its occupant leaves?
If the results of the orders were applied immediately, then the _ordering_
of the resolutions matters. With the new system, all the changes can
be considered as an ensemble.