Refactored the new turn creation in Turn#resolve!
This work is not over yet, but step by step we're improving the
logic, as shown, in parts, by the improved legibility.

First, splitting the tests in two helps distinguish the 2 important
things about the new turn: it's internal state, and especially its
status, and the state of its board.

The status **should not be defined**, since we want a new Turn
**that hasn't been started yet** – the decision to start it will
be made by another object (possibly `Game`). `Turn#resolve!` should
resolve all the turn's orders, update the turn's status, and prepare
a new turn, but not start it.

(Incidentally, the previous implementation of this behaviour was buggy,
which was spotted by the improved tests.)

As for the board, it should still be an updated version of the previous
turn's board, with the updates from the orders applied. This hasn't changed.