`Board#revise` now expects `Board::Update` object
It makes more sense to keep the logic of converting `Update` objects
to actual `Zone` objects in the `Board` space – the common namespace
was a strong hint.

I don't want to extract more logic from `Turn#resolve!`, though. I
feel like "merging" the updates is still the turn's responsibility.

In other words: `Turn` is the referee who revises the list of updates;
`Board` is simply in charge of placing the units tokens on the game
board according to the revised list of updates.

We could go a bit further, and introduce a `UpdatesList` object to
represent this list; this object could be in charge of merging its
own content (therefore taking over the role of the referee above), and/or
validate its own state, to ensure that no conflicting updates could
be passed to board. But it seems a bit overkill for now.