Small adjustments to `Turn#resolve!`
- The new turn is _created_ (i.e. persisted), and not simply built. This
  will make it possible to reuse it even if `#resolve!` was not called
  with a block – otherwise, the new turn would be lost.
- A finished turn cannot be resolved. This is because the resolution is
  not idempotent, since it updates the orders (and creates a new turn).

These two changes are related: because the method updates the orders,
it was not idempotent, and because it was not idempotent, saving the new
turn is necessary – otherwise, there would be no way to having it back,
if needed, since calling the method a second time is not an option. But
saving the turn makes the method even less idempotent (not that it matters
a lot).