Rename Zone to Hex
The naming was getting confusing… From a gaming, real-world perspective,
the board is divided into "squares", but because they are not
actual squares (this is not a Monopoly or chess board), we need a better
name.

Besides, from an application perspective, we have objects which _represent_
these "squares", with their characteristics (how many units are placed on
them, which player occupies them, etc.)

The word "zone" was used for both concepts, and this was getting confusing.
"Pass this method the zone as an argument": does the method expect the
name of a "square" (such as `:north`) or an actual `Zone` object?

Now, the squares are still called "zones", but their representations
are named `Hex` (because eventually, I'd like the board to be a hex map
of space, like in the Traveller RPG).

I hope that this change will ease the understanding of the code base,
and not make it even more confusing…