Let's move on to the email system!
OK, so I have a basic game engine – the rules are very minimal, and
even then they aren't fully implemented, but enough tinkering! Let's
add the e-mail system to send orders on top of this rudimentary
game, so that we can then move on to the _real_ stuff – parsing
natural(-ish) language, and then _generating_ some.