Renaming Hex to Tile
After a night's sleep, I realzied that _hex_ was as bad a term as _zone_.

The idea is to have objects that can be seen like physical things that
would be swapped on the physical board of the game. "Hex" describes
a shape, not an object – being so abstract, it could as well name
a real-life game component or be a colourful in-game term (I can imagine
a mech pilot saying "let's move to hex 13-37!" as much as "let's move to
zone 13-37!")

Anyway, naming IS hard.