OrdersMailbox finds the relevant game by itself
I didn't like the coupling that `Game.find_by_email_address` introduced
between the games and the emailing stuff. Granted, we may eventually
need for a game to have its own email address, and maybe generate it
itself, but at the moment the game engine could and should be usable
with any kind of interface – emails, but why not a web UI or the Rails
console.

So I want an interface between games and emails, and at the moment
`OrdersMailbox` seems a good fit – we'll see later if something else
becomes more suited.

Besides, there is no need to use a regexp to extract the id from the
email domain – since the id is the first part of the domain, a
simple casting does the trick, just like with URLs (e.g.
`/posts/123-post-title`)