Sending confirmation email when receiving an order
Let's do controversial stuff! I have NO PROBLEM with using a
callback in `Order` to send the confirmation email. I find this a
perfectly fine use case, no matter how many people complain about
callbacks. (And, yes, it does couple a model with a mailer. Relax, I know
what I'm doing.)

Adding a second email to the `OrdersMailer` lead me to reconsider what
the default parameters (see ActionMailer::Parameterized) should be. Just
like with controllers, always having an instance of the model that
the mail is about (Order for OrdersMailer, as for OrdersController) makes
a lot of sense; the other objects needed by `.error_no_participation`
(and only it!) are a bit harder to justify have as defaults, especially
since they'd be _optional_ defaults. So I removed them.