Retouching the notification and error emails
A few things about these two emails:

- As it happens, a mailer is either "parameterized" or not. You cannot
have callbacks that rely on `params` and _not_ use `with`. Well, I
guess that you _could_ with a lot of safeguards, but the fact that it's
not doable easily – i.e. that `params` is not always defined, even if
empty – tells me that it shouldn't be done.

Which makes sense: if the "parameterization" of an email is a way to
_systematically_ provide certain pieces of data, then said data should
_systematically_ be need. If a mailer has actions that don't require
it, maybe these actions aren't in the right mailer in the first place.

So I've split `OrdersMailer` in two: `OrdersMailer` is now meant for
emails relating to a specific `Order`, while `GamesMailer` is for
other emails that can be sent during a game – like an error notification,
for example. This is much clearer now.

I've also been a bit disappointed to realize that fixtures aren't easily
accessible from an email previewer. For some reason,
`ActiveRecord::FixtureSet` is not found (and I didn't bother to look
deeper). It's not that important because the emails are very minimal for
now, and so are their previewers.