DRY the logic of recipient address for orders
The rule is "orders@<game id>.<game domain>", and we need it twice:
once for the mailbox routing, and once to extract the game id when
processing the inbound mail.

Stealing the idea from GoRails' Chris Oliver
(https://gorails.com/episodes/routing-inbound-emails-actionmailbox-rails),
I've moved the regexp to `OrdersMailbox` and use it both there and in
the mailbox routing. It's neat.

(And I like the way the id is extracted in `OrdersMailbox#game`, even
if it is so arcane that I'll probably hate myself in two months. Or two
days.)