Gardening
Various useless stuff remove; no logic change, just a bit of code
cleanup.

There is one change worth mentioning, though: I've turned the test
inbound email in OrderMailboxTest into a fixture. It makes the test
easier to read, but I'm not sure that it's a good idea. The lack of
ERB in the file fixtures imposes to hardcode fixture ids. So, the
typical tradeoff of fixtures: less obvious boilerplate, but also
less context/more coupling. We'll see later if this one tradeoff wasn't
worth it.