Sliming a participation inbound mail
In other words, the ability to join a game by sending an email to
its address, with the simple subject "join".

This new feature helps me get a better understand of what we're trying
to achieve with this "chatbot via email" approach, and also reveals
issues with the current implementation.

"join" could be called a _command_ to the game (or its bot, aka its
_arbiter_). But "move" and "attack" could _also_ be commands – the only
difference being that these commands would lead to the reception of
_orders_, while other such as "join" would have other effects.

This similarity, and the emergence of the concept of "command", explains
the uncomfortable similarities (and the naming difficulties) between
`OrdersMailbox` and `CommandsMailbox`. So let's try to sort this out
before going any further.