Merging OrdersMailbox and CommandsMails
There were too many similarities between the two; now there's a single
mailbox, `ArbiterMailbox`. It is the mailbox for all _commands_ sent to
a game's arbiter – orders or others.

The implementation is still a very rough WIP, but at least the concepts
are emerging more clearly now:

- The arbiter expects _commands_.
- These commands are indicated by the **subject** of the email, while
  details, if needed, are passed in the body.
- Some commands may have preconditions (such as: an order cannot be
  given for a game in which the player is not participating).

We'll need a way to extract the command from the subject, maybe move
some validation to it, and overall this mailbox. But things are moving
in a sound direction!