Upgrading to edge Rails
It's not by accident if I've been listening to Red Fang's cover
of On The Edge, it was fate!

Or rather, I need to test the enqueuing of emails which use both
parameters and arguments, so I need this PR: https://github.com/rails/rails/pull/45752

It's a good opportunity to try out the bleeding edge stuff, though. Not
something I could have done on a real, in production app, but that's
the luxury of pet projects.