Introducing `Command` objects
I've been reluctant to do that (and I still am a bit, I must stay),
because they look awfully like _services_. And yet, it does seem
reasonable to wrap a complex-y process in an object, isolated from
the UI (like a mailbox, for example).

The good side effect of this new set of (non-AR) models is that it
provides an object to take care of the parsing of commands sent by email.