The single mailbox delegates to Command
Things are getting clear again. The mailbox's job is simple and clearly
defined: when a command is received by email, the mailbox ensures that
the email is valid, and then fetches the command, fully formed from
the email (the subject is parsed to identify the required command, and
the details, if necessary, are extracted from the email body or headers).

If the command succeed, nothing special happens (the inbound email is
marked as processed). If not, a notification is sent to the sender.

Because the mailbox doesn't know the details of the command, nor the
reasons for its failure, it needs to pass the details of the errors
to the mailer. This is a bit cumbersome, but we'll live.