WIP: removing the commands validation
These validations where actually deguised validations of the
underpinned models of the commands: `Order` for `ReceiveOrder`,
`Participation` for `JoinGame`, etc.

Commands are basically controller action at this point, and if we had
controller actions, then there would only be models validations. So
all this makes sense, and goes toward making command an even thinner
layer.

However, it does makes _returning info_ (the reasons for the failure)
to the upper layer (the mailbox) even harder.

We could go with a custom serializer for `Errors`, and maybe even a
custom class too, but this would only make this layer large again. So
instead let's reduce the `Errors` object to a simpler hash, and let
the mailer deal with this "lossly compressed" info.