WIP: removing the Command objects
Back to the start… I like the _idea_ of having a generic abstraction
for a "command", but implementing it, via simple or less simple POROs,
is not satisfying. All abstractions are leaky, and here the leaks
seem to be around proving feedback when something goes wrong.

This very simple approach goes back to keeping all "command resolution"
logic in the Mailbox object. It does make it less testable at the unit
level, and definitely not usable with another I/O mechanism (a bot/arbiter
via SMS for example), but we're not there yet, so maybe it's a good
compromise.

(However, the way the notification email is sent when a command fails is
still shitty.)