Removed Commands completely
Even further in the trimming down of all I did before. Since "commands"
seem to always be CRUD actions on core models ("join a game" would
be "participations#create" in a web app, "receive this order" would be
"orders#create", etc.), and more specifically Create actions,
let's rely solely on these code models.

Now the mailbox's role is basically to identify the right "commanded"
model to create. All controls (aka validations) can and should be on
the model; if saving doesn't work, a bounce email can be sent with the
errors from the model.

(Since there is only one email, turning the different errors into a
common email content could be tricky, but 1/ it will be the mailer's
responsibility and 2/ this is where natural language generation could
be fun.)