Pushed more "command" logic to the core models
We're getting to a very lean ArbiterMailbox, which is basically only
tasked with creating and save a single record; the record's class
(and attributes) depends on the "command", identified through the
mail's subject.

The validity checks (e.g. "can this player join this game?") are handled
by simple AR validations, as are the collateral creations (e.g. creating
a new `Player` record when a new user wants to join a game).

Basic AR features like nested attributes and validations are enough to
provide all the functionalities of a "command". The final piece of the
puzzle was: _how to let the user what went wrong, if anything goes wrong?_.

Because the potentials errors are represented by an `Errors` object,
we could have a mailer that uses this object to unpack the details
of "what went wrong", and builds the content of its email from that.

However, when bouncing an incoming email, the email sent along the
bounce (`bounce_with MyMailer.whatever`_) is sent asynchronously. Which
means that whatever params and args are passed, they must be serializable
by ActiveJob. Which may not be the case here, since the `Errors` object
 could include references to non-persisted AR objects. (For example,
if the creation of a `Participation` failed because its `Player`,
which was created along, was itself invalid. In this case, the `Errors`
object will include the `Player` object, and this object will have no
ID, so won't be serializable.).

To circumvent this issue, instead of passing a raw `Errors` object,
we pass a simplified Hash which conveys all the important info (the
error types for each attribute), reduced to serializable symbols, arrays
and hashes.

Fun thing: to have this simplified hash, we use a Refinement to Errors.
This allows us to have a nice API `errors.types_only`) without
monkey-patching Rails.

However, there is _still_ a possible serialization issues. `ArbiterMailer`
also accepts two parameters: a `Game` and a `Player` objects. Once again,
a failed operation could result in new, but invalid and unpersisted, records
(e.g. the `Player` object creating when join a game.) This is the same issue
as when passing errors to the mailer, but this time, we cannot avoid
it by transforming the problematic object into something else. So, instead,
I've had to replace the `player` param by a `player_email_address`, which
should always by serializable. But it's a bit sad.