Bouncing with a sync email
When a "command" fails, instead of bouncing with an async
email (which is what the original `#bounce_with` does), let's
bounce with a "sync" email, by refining `ActionMailbox::Base` so
that `#bounce_with` accepts a `deliver_now` argument.

And now, since we're not deferring the sending of the bounce email,
we don't go through ActiveJob, so we don't need to serialize the mailer's
params and arguments. Which simplifies a lot of things, in exchange for
a small surprise.

(And it's honestly that big of a surprise, since the issue of unserializable
arguments to mailers is not new.)

This also means that the mailer method can receive a whole AR object, with
its validation state, instead of a simplified hash of the validation errors.