CommandsMailer
It handles all the emails for a _failed_ command, generating the
relevant message based on the `Errors` of the "commanded" object. Which
implies that a command is always, at its core, _doing something to a
model_ (like creating it), so the command's errors are actually the
model's errors. This assumption may not hold, in which case this approach
would need to be changed.

Also note that the generation of the email message is left to the mailer,
and not to the model (i.e. validation messages such as `Errors#messages`
are not used). This is to allow eventual calls to a natural language
generation system, instead of hard coded messages.