Move the "order from text" back to the mailbox
We will eventually need an object to handle the NLP that will turn an
incoming email into a new `Order` object. However, at the moment, I find
this responsibility better handed off to the `CommandsMailbox`. It is
definitely not SRP-compliant, but since the mailbox already prepares
the resources for other commands (such as the `Participation` if the
incoming email is a request to join a game), let's at least have a
coherent architecture – even if it makes the mailbox object do too
many things for a while.